+++ CHANGES.md
@@ -4,6 +4,7 @@
  * [METAMODEL-138] - Allow empty characteres before AS keyword.
  * [METAMODEL-141] - Improved mapping of ColumnType to SQL data types for Oracle, SQL Server, MySQL, DB2 and PostgreSQL
  * [METAMODEL-142] - Ensured that JDBC schema refreshes in an UpdateScript is using same Connection/Transaction as rest of operations
+ * [METAMODEL-133] - Improved query parser support for multiple JOINs in same query.
 
 ### Apache MetaModel 4.3.3